casted a few UI elements to ints so they cleaner

+++ MonoGame Dynamics Final Project/MonoGame Dynamics Final Project/Game1.cs
@@ -477,6 +477,7 @@ namespace MonoGame_Dynamics_Final_Project
                             {
                                 if (Enemywave[i].enemyType == "stingRay1" && playGame == true)
                                 {
+
                                     StingrayParticles.RemoveAt(StingrayParticles.Count - 1);
                                 }
                                 if (Enemywave[i].enemyType == "stingRay2" && playGame == true)
@@ -1065,11 +1066,11 @@ namespace MonoGame_Dynamics_Final_Project
                     spriteBatch.Begin();
 
                     spriteBatch.Draw(health, healthRect, Color.White);
-                    spriteBatch.DrawString(menuFont, "hp:" + playerShip.Health + " / " + playerShip.MaxHealth, new Vector2(110.0f, GraphicsDevice.Viewport.Height - 47.0f), Color.White, 0.08f);
+                    spriteBatch.DrawString(menuFont, "hp:" + (int)playerShip.Health + " / " + playerShip.MaxHealth, new Vector2(110.0f, GraphicsDevice.Viewport.Height - 47.0f), Color.White, 0.08f);
                     spriteBatch.Draw(xp, xpRect, Color.White);
-                    spriteBatch.DrawString(menuFont, "xp:" + playerShip.Experience + "/" + playerShip.ExperienceToNextLevel, new Vector2(110.0f, GraphicsDevice.Viewport.Height - 27.0f), Color.White, 0.08f);
+                    spriteBatch.DrawString(menuFont, "xp:" + (int)playerShip.Experience + "/" + playerShip.ExperienceToNextLevel, new Vector2(110.0f, GraphicsDevice.Viewport.Height - 27.0f), Color.White, 0.08f);
                     spriteBatch.DrawString(menuFont, "Lvl:" + playerShip.Level, new Vector2(GraphicsDevice.Viewport.Width - 200.0f, 50.0f), Color.White, 0.25f);
-                    spriteBatch.DrawString(menuFont, "Score:" + score, new Vector2(0.0f, 50.0f), Color.White, 0.0f,Vector2.Zero,0.25f,SpriteEffects.None,0.0f);
+                    spriteBatch.DrawString(menuFont, "Score:" + (int)score, new Vector2(0.0f, 50.0f), Color.White, 0.0f,Vector2.Zero,0.25f,SpriteEffects.None,0.0f);
                     spriteBatch.DrawString(menuFont, "Primary:" + playerShip.PrimaryType, new Vector2(100.0f, GraphicsDevice.Viewport.Height - 80.0f), Color.White, 0.0f, Vector2.Zero, 0.15f, SpriteEffects.None, 0.0f);
                     spriteBatch.DrawString(menuFont, "Secondary:" + playerShip.SecondaryType, new Vector2(GraphicsDevice.Viewport.Width - 550.0f, GraphicsDevice.Viewport.Height - 80.0f), Color.White, 0.0f,Vector2.Zero,0.15f,SpriteEffects.None,0.0f);
